python_project_internship/project/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
flask
textblob
speechrecognition
pyttsx3
pyaudio
requests
google-generativeai
openai


